embeded native dlls into managed wrapper dll minor change to project's config

+++ RenderCore/RenderElement.cpp
@@ -9,7 +9,7 @@ vector<wstring> DrawableHelper::typeMap;
 DrawableHelper::DrawableHelper(const wstring& name)
 {
 	typeMap.push_back(name);
-	id = typeMap.size() - 1;
+	id = static_cast<uint32_t>(typeMap.size() - 1);
 #ifdef _DEBUG
 	printf("@@@@regist Drawable [%ls] -> %2d\n", name.c_str(), id);
 #endif
